backend/requirements.txt
fastapi
uvicorn
websockets
pandas
scikit-learn
requests
python-dotenv
xgboost